Add files via upload
In this simple study I accessed the variables from some youtube videos through its API. By comparing the variables from the videos, we can see that 'unpopular' videos in this small dataset tend to engage its viewers more (cause them to react with likes, dislikes or comments). We can also observe that these unpopular videos don't really get as many views as the popular ones, therefore the actual number of reactions they get is often not as high as the number of reactions a popular video would get. Although, one of the unpopular videos caused almost as many reactions as some of the popular ones.